More work on training input
