--- python/biomedicus/sentences/__main__.py
@@ -1,27 +1,0 @@
-# Copyright 2019 Regents of the University of Minnesota.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-import logging
-
-from biomedicus.sentences.commands import SentenceCommands, create_parser
-
-logging.basicConfig(level=logging.INFO)
-
-
-def main(args=None):
-    parser = create_parser()
-    commands = parser.parse_args(args=args, namespace=SentenceCommands())
-    commands.func(commands)
-
-
-main()
