+++ python/biomedicus/sentences/models.py
@@ -109,45 +109,6 @@ class BiLSTMSentenceModel(SentenceModel):
 
     """
 
-    def __init__(self, vocabulary: Vocabulary, args: Namespace):
-        super().__init__()
-        self.vocabulary = vocabulary
-        self.labels = self.vocabulary.labels
-        self.chars = self.vocabulary.characters
-
-        self.words = self.vocabulary.words
-        self.word_embeddings = self.vocabulary.word_vectors
-        self.dim_word = args.dim_word or self.vocabulary.get_word_dimension()
-
-        self.sequence_length = args.sequence_length
-
-        self.use_chars = args.use_chars
-
-        self.word_length = args.word_length
-        self.dim_char = args.dim_char
-
-        self.char_mode = args.char_mode
-        self.char_cnn_filters = args.char_cnn_filters
-        self.char_cnn_kernel_size = args.char_cnn_kernel_size
-        self.char_lstm_hidden_size = args.char_lstm_hidden_size
-
-        self.use_words = args.use_words
-
-        self.lstm_hidden_size = args.lstm_hidden_size
-
-        self.dropout = args.dropout
-        self.recurrent_dropout = args.recurrent_dropout
-
-        self.concatenate_words_chars = args.concatenate_word_chars
-
-        self._model = self._build_model()
-
-    def _build_model(self) -> tf.keras.models.Model:
-
-
-    def build_layers(self):
-
-
     @property
     def model(self) -> tf.keras.models.Model:
         return self._model
@@ -155,18 +116,7 @@ class BiLSTMSentenceModel(SentenceModel):
     def compile_model(self,
                       optimizer: Union[AnyStr, tf.keras.optimizers.Optimizer]):
         self.model.compile(optimizer=optimizer,
-                           sample_weight_mode='temporal',
-                           loss={
-                               'logits': 'binary_crossentropy',
-                               'contextual_word_representation': None
-                           },
-                           weighted_metrics={
-                               'logits': ['binary_accuracy']
-                           },
-                           loss_weights={
-                               'logits': 1.,
-                               'contextual_word_representation': 0.
-                           })
+                           )
 
     def get_config(self):
         config = dict(self.__dict__)
@@ -177,226 +127,6 @@ class BiLSTMSentenceModel(SentenceModel):
         config.pop('word_embeddings')
         config.pop('_model')
         return config
-
-
-class InputMapper(metaclass=ABCMeta):
-
-    @abstractmethod
-    def map_input(self,
-                  txt_tokens: Iterable[Tuple[AnyStr, List[Token]]],
-                  include_labels: bool) -> Tuple:
-        raise NotImplementedError
-
-
-class DeepMapper(InputMapper):
-    def __init__(self, vocabulary: Vocabulary, args: Namespace):
-        self.vocabulary = vocabulary
-
-        self.sequence_length = args.sequence_length
-
-        self.use_chars = args.use_chars
-
-        self.word_length = args.word_length
-
-        self.use_token_boundaries = args.use_token_boundaries
-        self.use_neighbor_boundaries = args.use_neighbor_boundaries
-        self.use_sequence_boundaries = args.use_sequence_boundaries
-
-        self.use_words = args.use_words
-
-    def map_input(self,
-                  txt_tokens: Iterable[Tuple[AnyStr, List[Token]]],
-                  include_labels: bool) -> Tuple:
-        generator = InputGenerator(txt_tokens,
-                                   vocabulary=self.vocabulary,
-                                   sequence_length=self.sequence_length,
-                                   word_length=self.word_length,
-                                   use_chars=self.use_chars,
-                                   use_token_boundaries=self.use_token_boundaries,
-                                   use_neighbor_boundaries=self.use_neighbor_boundaries,
-                                   use_sequence_boundaries=self.use_sequence_boundaries,
-                                   include_labels=include_labels,
-                                   use_words=self.use_words)
-        inputs = next(iter(generator))
-        return inputs
-
-
-class InputGenerator(TokenSequenceGenerator):
-    """Turns a iterable of (text, List[Token]) tuples into an iterable of input or (input, label)
-    tuples
-    """
-
-    def __init__(self,
-                 input_source: Iterable[Tuple[AnyStr, List[Token]]],
-                 vocabulary: Vocabulary,
-                 sequence_length: int,
-                 word_length: int,
-                 use_chars: bool,
-                 use_words: bool,
-                 use_token_boundaries: bool,
-                 use_neighbor_boundaries: bool,
-                 use_sequence_boundaries: bool,
-                 include_labels: bool = True,
-                 batch_size: int = -1):
-        super().__init__(input_source, batch_size, sequence_length)
-        self.vocabulary = vocabulary
-
-        self.word_length = word_length
-
-        self.use_chars = use_chars
-        self.use_words = use_words
-        self.use_token_boundaries = use_token_boundaries
-        self.use_neighbor_boundaries = use_neighbor_boundaries
-        self.use_sequence_boundaries = use_sequence_boundaries
-
-        self.batch_count = 0
-        self.chars = []
-        self.words = []
-        self.labels = []
-        self.weights = []
-
-        self.sequence_count = 0
-        self.segment_chars = []
-        self.segment_words = []
-        self.segment_labels = []
-        self.segment_weights = []
-
-        self.include_labels = include_labels
-
-        self.class_counts = {'B': 0, 'I': 0, 'O': 0}
-
-    def _handle_token(self):
-        """Turns the token into its character ids and word id and adds label to the current
-        sequence.
-        """
-        if self.use_chars:
-            chars = self.get_chars()
-            self.segment_chars.append(chars)
-
-        if self.use_words:
-            word_id = self.vocabulary.get_word_id(
-                self.txt[self.current.start_index:self.current.end_index],
-                is_identifier=self.current.is_identifier)
-            self.segment_words.append(word_id)
-
-        if self.include_labels:
-            label_id = self.vocabulary._label_to_id[self.current.label]
-            self.class_counts[self.current.label] += 1
-            self.segment_labels.append(label_id)
-            if self.current.label == 'O':
-                self.segment_weights.append(0.)
-            else:
-                self.segment_weights.append(1.)
-        else:
-            self.segment_weights.append(1.)
-
-    def _finish_sequence(self):
-        if self.use_chars:
-            self.chars.append(pad_to_length(self.segment_chars,
-                                            length=self.word_length,
-                                            value=0))
-            self.segment_chars = []
-
-        if self.use_words:
-            self.words.append(self.segment_words)
-            self.segment_words = []
-
-        self.weights.append(self.segment_weights)
-        self.segment_weights = []
-
-        if self.include_labels:
-            self.labels.append(self.segment_labels)
-            self.segment_labels = []
-
-    def _batch(self):
-        inputs = {}
-
-        if self.use_chars:
-            inputs['char_input'] = pad_to_length(self.chars,
-                                                 length=self.sequence_length,
-                                                 value=numpy.zeros(
-                                                     self.word_length))
-            self.chars = []
-        if self.use_words:
-            inputs['word_input'] = pad_to_length(
-                self.words,
-                length=self.sequence_length,
-                value=0
-            )
-            self.words = []
-
-        class_counts = self.class_counts
-        self.class_counts = {'B': 0, 'I': 0, 'O': 0}
-        weights = pad_to_length(
-            self.weights,
-            length=self.sequence_length,
-            value=0.
-        )
-        self.weights = []
-
-        if self.include_labels:
-            labels = pad_to_length(
-                self.labels,
-                length=self.sequence_length,
-                value=0
-            )
-            self.labels = []
-            labels = labels[:, :, numpy.newaxis]
-
-            return inputs, class_counts, weights, labels
-        else:
-            return inputs, class_counts, weights
-
-    def get_chars(self) -> List[int]:
-        begin = self.current.start_index
-        end = self.current.end_index
-
-        all_chars = []
-
-        if self.prev is None:
-            prev_end = max(0, begin - 7)
-        else:
-            prev_end = max(self.prev.end_index, begin - 7)
-
-            if self.use_neighbor_boundaries:
-                all_chars.append('PREV_TOKEN')
-
-        pre = self.txt[prev_end:begin]
-        all_chars += list(pre)
-
-        if self.use_sequence_boundaries:
-            if self.prev is None or self.current.segment != self.prev.segment:
-                all_chars.append('SEGMENT_BEGIN')
-
-        if self.use_token_boundaries:
-            all_chars.append('TOKEN_BEGIN')
-
-        if self.current.is_identifier:
-            all_chars.append('IDENTIFIER')
-        else:
-            token_txt = self.txt[begin:end]
-            all_chars += list(token_txt)
-
-        if self.use_token_boundaries:
-            all_chars.append('TOKEN_END')
-
-        if self.use_sequence_boundaries:
-            if self.next is None or self.current.segment != self.next.segment:
-                all_chars.append('SEGMENT_END')
-
-        if self.next is not None:
-            next_begin = min(self.next.start_index, end + 7)
-        else:
-            next_begin = min(len(self.txt), end + 7)
-
-        post = self.txt[end:next_begin]
-        all_chars += list(post)
-
-        if self.use_neighbor_boundaries and self.next is not None:
-            all_chars.append('NEXT_TOKEN')
-
-        # TODO: Look into doing id lookup prior to appending rather than after
-        return [self.vocabulary.get_character_id(i) for i in all_chars]
 
 
 def ensemble_hparams_parser():
